requirements.txt
PyYAML==3.12
requests==2.18.4
GitPython==2.1.9
distro==1.3.0
packaging==19.0
